mypy limpio: 0 errores en 73 ficheros
Los 42 que quedaban. `ruff check`, `ruff format`, mypy y 997 pruebas, todo en
verde a la vez.

CUATRO QUE PODIAN REVENTAR EN CALIENTE, los cuatro por un None no contemplado:

- `_aware(segment.played_at)` con la fecha vacia: AttributeError al pedirle
  `tzinfo` a un None. Un segmento sin fecha no se puede situar en ningun ciclo
  de entrenamiento, asi que se salta.
- `default_setup(training_target(...))`: `training_target` devuelve None para
  un tipo de entrenamiento que no conoce. Se usa el mismo respaldo que ya habia
  dos lineas mas arriba.
- `weeks_owned(desde, hasta)` con la fecha de cierre vacia. `desde` ya estaba
  protegido; `hasta` no.
- `rng.poisson(media)` en el simulador: de cada par, quien llama pone el numero
  fijo O la media, nunca los dos. Ese contrato ahora esta escrito y, si algun
  dia llegan los dos vacios, salta un error que lo dice en vez de uno de numpy.

DOS FIRMAS QUE MENTIAN:

- `training_efficiency_pct(stamina_share: int)` recibe un float, y siempre lo
  recibio: `TrainingSetup.stamina_share` es float.
- `promotion_cost=conv(CONSTANTE)`: `conv` solo devuelve None si RECIBE None,
  asi que obligaba a tratar como opcional algo que nunca falta.

El resto es deuda de anotacion: diez firmas sin tipos, dieciseis `return` que
devolvian `Any` desde funciones que prometian otra cosa --casi todos
`session.scalar()`, que esta tipado como `Any`--, y un par de variables de
bucle reutilizadas con dos tipos distintos.

Un tira y afloja entre las dos herramientas: `dict()` de una `Row` de
SQLAlchemy funciona por casualidad y mypy lo rechaza, pero la version con
comprension la rechaza ruff. La salida es `.tuples()`, que convierte de verdad.

Y dos errores mios por el camino, cazados por las propias herramientas: anote
`str | None` donde la funcion devuelve `int | None`, y deje tres funciones
asignando a una variable sin devolverla.

+++ backend/app/domain/engines/staff_effects.py
@@ -128,7 +128,7 @@ def current_injury_risk_pct(assistant_combined_level: int, doctor_level: int) ->
     combinado) + doctor nivel 5 = 0.275 lesiones/partido (27.5%)."""
     assistant = assistant_trainer_effect(assistant_combined_level)["injuryRiskPp"]
     doctor = doctor_effect(doctor_level)["injuryRiskReductionPp"]
-    return round(BASE_INJURY_RISK_PCT + assistant - doctor, 2)
+    return float(round(BASE_INJURY_RISK_PCT + assistant - doctor, 2))
 
 
 # HL-2xx: qué función de efecto corresponde a cada columna de StaffSnapshot
